Separate ApiGateway and ApiGatewayWebSockets from Plugin
