+++ src/index.js
@@ -1,37 +1,13 @@
 'use strict';
 
-// Node dependencies
-const fs = require('fs');
-const path = require('path');
-const { performance, PerformanceObserver } = require('perf_hooks');
 const { exec } = require('child_process');
-
-// External dependencies
-const hapi = require('@hapi/hapi');
-const h2o2 = require('@hapi/h2o2');
-
-// Internal lib
+const ApiGateway = require('./ApiGateway');
+const ApiGatewayWebSocket = require('./ApiGatewayWebSocket');
 const debugLog = require('./debugLog');
-const jsonPath = require('./jsonPath');
-const createLambdaContext = require('./createLambdaContext');
-const createVelocityContext = require('./createVelocityContext');
-const createLambdaProxyContext = require('./createLambdaProxyContext');
-const renderVelocityTemplateObject = require('./renderVelocityTemplateObject');
-const createAuthScheme = require('./createAuthScheme');
 const functionHelper = require('./functionHelper');
-const Endpoint = require('./Endpoint');
-const parseResources = require('./parseResources');
-const { createDefaultApiKey, detectEncoding, getUniqueId } = require('./utils');
-const authFunctionNameExtractor = require('./authFunctionNameExtractor');
-const requestBodyValidator = require('./requestBodyValidator');
-const wsHelpers = require('./websocketHelpers');
+const { createDefaultApiKey } = require('./utils');
 
-/*
-  I'm against monolithic code like this file
-  but splitting it induces unneeded complexity.
-*/
-class Offline {
-
+module.exports = class Offline {
   constructor(serverless, options) {
     this.serverless = serverless;
     this.service = serverless.service;
@@ -39,8 +15,6 @@ class Offline {
     this.options = options;
     this.exitCode = 0;
     this.clients = new Map();
-    this.wsActions = {};
-    this.websocketsApiRouteSelectionExpression = serverless.service.provider.websocketsApiRouteSelectionExpression || '$request.body.action';
 
     this.commands = {
       offline: {
@@ -178,7 +152,8 @@ class Offline {
     process.env.IS_OFFLINE = true;
 
     return Promise.resolve(this._buildServer())
-      .then(() => this._listen())
+      .then(() => this.apiGateway._listen())
+      .then(() => this.apiGatewayWebSocket._listen())
       .then(() => this.options.exec ? this._executeShellScript() : this._listenForTermination());
   }
 
@@ -241,19 +216,20 @@ class Offline {
   }
 
   _buildServer() {
-    // Maps a request id to the request's state (done: bool, timeout: timer)
-    this.requests = {};
-
-    // Methods
     this._setOptions(); // Will create meaningful options from cli options
     this._storeOriginalEnvironment(); // stores the original process.env for assigning upon invoking the handlers
-    this._createServer(); // Hapijs boot
-    this._createRoutes(); // API  Gateway emulation
-    this._createResourceRoutes(); // HTTP Proxy defined in Resource
-    this._create404Route(); // Not found handling
-    this._createWebSocket();
 
-    return this.server;
+    this.apiGateway = new ApiGateway(this.serverless, this.options, this.velocityContextOptions);
+    this.apiGatewayWebSocket = new ApiGatewayWebSocket(this.serverless, this.options);
+
+    const server = this.apiGateway._createServer();
+    this.apiGatewayWebSocket._createWebSocket();
+
+    this.apiGateway._createRoutes(); // API  Gateway emulation
+    this.apiGateway._createResourceRoutes(); // HTTP Proxy defined in Resource
+    this.apiGateway._create404Route(); // Not found handling
+
+    return server;
   }
 
   _storeOriginalEnvironment() {
@@ -330,1236 +306,14 @@ class Offline {
     debugLog('options:', this.options);
   }
 
-  _createServer() {
-    const serverOptions = {
-      host: this.options.host,
-      port: this.options.port,
-      router: {
-        stripTrailingSlash: !this.options.preserveTrailingSlash, // removes trailing slashes on incoming paths.
-      },
-    };
-
-    const httpsDir = this.options.httpsProtocol;
-
-    // HTTPS support
-    if (typeof httpsDir === 'string' && httpsDir.length > 0) {
-      serverOptions.tls = {
-        key: fs.readFileSync(path.resolve(httpsDir, 'key.pem'), 'ascii'),
-        cert: fs.readFileSync(path.resolve(httpsDir, 'cert.pem'), 'ascii'),
-      };
-    }
-
-    serverOptions.state = this.options.enforceSecureCookies ? {
-      isHttpOnly: true,
-      isSecure: true,
-      isSameSite: false,
-    } : {
-      isHttpOnly: false,
-      isSecure: false,
-      isSameSite: false,
-    };
-
-    // Hapijs server creation
-    this.server = hapi.server(serverOptions);
-
-    this.server.register(h2o2).catch(err => err && this.serverlessLog(err));
-
-    // Enable CORS preflight response
-    this.server.ext('onPreResponse', (request, h) => {
-      if (request.headers.origin) {
-        const response = request.response.isBoom ? request.response.output : request.response;
-
-        response.headers['access-control-allow-origin'] = request.headers.origin;
-        response.headers['access-control-allow-credentials'] = 'true';
-
-        if (request.method === 'options') {
-          response.statusCode = 200;
-          response.headers['access-control-expose-headers'] = 'content-type, content-length, etag';
-          response.headers['access-control-max-age'] = 60 * 10;
-
-          if (request.headers['access-control-request-headers']) {
-            response.headers['access-control-allow-headers'] = request.headers['access-control-request-headers'];
-          }
-
-          if (request.headers['access-control-request-method']) {
-            response.headers['access-control-allow-methods'] = request.headers['access-control-request-method'];
-          }
-        }
-      }
-
-      return h.continue;
-    });
-  }
-
-  _createWebSocket() {
-    // start COPY PASTE FROM HTTP SERVER CODE
-    const serverOptions = {
-      host: this.options.host,
-      port: this.options.wsPort,
-      router: {
-        stripTrailingSlash: !this.options.preserveTrailingSlash, // removes trailing slashes on incoming paths.
-      },
-    };
-
-    const httpsDir = this.options.httpsProtocol;
-
-    // HTTPS support
-    if (typeof httpsDir === 'string' && httpsDir.length > 0) {
-      serverOptions.tls = {
-        key: fs.readFileSync(path.resolve(httpsDir, 'key.pem'), 'ascii'),
-        cert: fs.readFileSync(path.resolve(httpsDir, 'cert.pem'), 'ascii'),
-      };
-    }
-
-    serverOptions.state = this.options.enforceSecureCookies ? {
-      isHttpOnly: true,
-      isSecure: true,
-      isSameSite: false,
-    } : {
-      isHttpOnly: false,
-      isSecure: false,
-      isSameSite: false,
-    };
-
-    // Hapijs server creation
-    this.wsServer = hapi.server(serverOptions);
-
-    this.wsServer.register(h2o2).catch(err => err && this.serverlessLog(err));
-
-    // Enable CORS preflight response
-    this.wsServer.ext('onPreResponse', (request, h) => {
-      if (request.headers.origin) {
-        const response = request.response.isBoom ? request.response.output : request.response;
-
-        response.headers['access-control-allow-origin'] = request.headers.origin;
-        response.headers['access-control-allow-credentials'] = 'true';
-
-        if (request.method === 'options') {
-          response.statusCode = 200;
-          response.headers['access-control-expose-headers'] = 'content-type, content-length, etag';
-          response.headers['access-control-max-age'] = 60 * 10;
-
-          if (request.headers['access-control-request-headers']) {
-            response.headers['access-control-allow-headers'] = request.headers['access-control-request-headers'];
-          }
-
-          if (request.headers['access-control-request-method']) {
-            response.headers['access-control-allow-methods'] = request.headers['access-control-request-method'];
-          }
-        }
-      }
-
-      return h.continue;
-    });
-    // end COPY PASTE FROM HTTP SERVER CODE
-
-    this.wsServer.register(require('hapi-plugin-websocket')).catch(err => err && this.serverlessLog(err));
-
-    const doAction = (ws, connectionId, name, event, context, doDeafultAction/* , onError */) => {
-      const sendError = err => {
-        if (ws.readyState === /* OPEN */1) ws.send(JSON.stringify({ message:'Internal server error', connectionId, requestId:'1234567890' }));
-        debugLog(`Error in handler of action ${action}`, err);
-      };
-      let action = this.wsActions[name];
-      if (!action && doDeafultAction) action = this.wsActions.$default;
-      if (!action) return;
-      let p = null;
-      try {
-        p = action.handler(event, context, err => {
-          if (!err) return;
-          sendError(err);
-        });
-      }
-      catch (err) {
-        sendError(err);
-      }
-
-      if (p) {
-        p.catch(err => {
-          sendError(err);
-        });
-      }
-    };
-
-    this.wsServer.route({
-      method: 'POST',
-      path: '/',
-      config: {
-        payload: { output: 'data', parse: true, allow: 'application/json' },
-        plugins: {
-          websocket: {
-            only: true,
-            initially: false,
-            connect: ({ ws, req }) => {
-              const parseQuery = queryString => {
-                const query = {}; const parts = queryString.split('?');
-                if (parts.length < 2) return {};
-                const pairs = parts[1].split('&');
-                pairs.forEach(pair => {
-                  const kv = pair.split('=');
-                  query[decodeURIComponent(kv[0])] = decodeURIComponent(kv[1] || '');
-                });
-
-                return query;
-              };
-
-              const queryStringParameters = parseQuery(req.url);
-              const connection = { connectionId:getUniqueId(), connectionTime:Date.now() };
-              debugLog(`connect:${connection.connectionId}`);
-
-              this.clients.set(ws, connection);
-              let event = wsHelpers.createConnectEvent('$connect', 'CONNECT', connection, this.options);
-              if (Object.keys(queryStringParameters).length > 0) event = { queryStringParameters, ...event };
-              const context = wsHelpers.createContext('$connect');
-
-              doAction(ws, connection.connectionId, '$connect', event, context);
-            },
-            disconnect: ({ ws }) => {
-              const connection = this.clients.get(ws);
-              debugLog(`disconnect:${connection.connectionId}`);
-              this.clients.delete(ws);
-              const event = wsHelpers.createDisconnectEvent('$disconnect', 'DISCONNECT', connection, this.options);
-              const context = wsHelpers.createContext('$disconnect');
-
-              doAction(ws, connection.connectionId, '$disconnect', event, context);
-            },
-          },
-        },
-      },
-      
-      handler: (request, h) => {
-        const { initially, ws } = request.websocket();
-        if (!request.payload || initially) return h.response().code(204);
-        const connection = this.clients.get(ws);
-        let actionName = null;
-        if (this.websocketsApiRouteSelectionExpression.startsWith('$request.body.')) {
-          actionName = request.payload;
-          if (typeof actionName === 'object') {
-            this.websocketsApiRouteSelectionExpression.replace('$request.body.', '').split('.').forEach(key => {
-              if (actionName) actionName = actionName[key];
-            });
-          }
-          else actionName = null;
-        }
-        if (typeof actionName !== 'string') actionName = null;
-        const action = actionName || '$default';
-        debugLog(`action:${action} on connection=${connection.connectionId}`);
-        const event = wsHelpers.createEvent(action, 'MESSAGE', connection, request.payload, this.options);
-        const context = wsHelpers.createContext(action);
-
-        doAction(ws, connection.connectionId, action, event, context, true);
-
-        return h.response().code(204);
-      },
-    });
-
-    this.wsServer.route({
-      method: 'GET',
-      path: '/{path*}',
-      handler: (request, h) => h.response().code(426),
-    });
-
-    this.wsServer.route({
-      method: 'POST',
-      path: '/@connections/{connectionId}',
-      config: { payload: { parse: false } },
-      handler: (request, h) => {
-        debugLog(`got POST to ${request.url}`);
-        const getByConnectionId = (map, searchValue) => {
-          for (const [key, connection] of map.entries()) {
-            if (connection.connectionId === searchValue) return key;
-          }
-
-          return undefined;
-        };
-
-        const ws = getByConnectionId(this.clients, request.params.connectionId);
-        if (!ws) return h.response().code(410);
-        if (!request.payload) return '';
-        ws.send(request.payload.toString());
-        // console.log(`sent "${request.payload.toString().substring}" to ${request.params.connectionId}`);
-        debugLog(`sent data to connection:${request.params.connectionId}`);
-
-        return '';
-      },
-    });
-  }
-
-  _createWsAction(fun, funName, servicePath, funOptions, event) {
-    let handler; // The lambda function
-    Object.assign(process.env, this.originalEnvironment);
-
-    try {
-      if (this.options.noEnvironment) {
-        // This evict errors in server when we use aws services like ssm
-        const baseEnvironment = {
-          AWS_REGION: 'dev',
-        };
-        if (!process.env.AWS_PROFILE) {
-          baseEnvironment.AWS_ACCESS_KEY_ID = 'dev';
-          baseEnvironment.AWS_SECRET_ACCESS_KEY = 'dev';
-        }
-
-        process.env = Object.assign(baseEnvironment, process.env);
-      }
-      else {
-        Object.assign(
-          process.env,
-          { AWS_REGION: this.service.provider.region },
-          this.service.provider.environment,
-          this.service.functions[funName].environment
-        );
-      }
-      process.env._HANDLER = fun.handler;
-      handler = functionHelper.createHandler(funOptions, this.options);
-    }
-    catch (err) {
-      return this.serverlessLog(`Error while loading ${funName}`, err);
-    }
-
-    const actionName = event.websocket.route;
-    const action = { funName, fun, funOptions, servicePath, handler };
-    this.wsActions[actionName] = action;
-    this.serverlessLog(`Action '${event.websocket.route}'`);
-  }
-
-  _createRoutes() {
-    let serviceRuntime = this.service.provider.runtime;
-    const defaultContentType = 'application/json';
-    const apiKeys = this.service.provider.apiKeys;
-    const protectedRoutes = [];
-
-    if (!serviceRuntime) {
-      throw new Error('Missing required property "runtime" for provider.');
-    }
-
-    if (typeof serviceRuntime !== 'string') {
-      throw new Error('Provider configuration property "runtime" wasn\'t a string.');
-    }
-
-    if (serviceRuntime === 'provided') {
-      if (this.options.providedRuntime) {
-        serviceRuntime = this.options.providedRuntime;
-      }
-      else {
-        throw new Error('Runtime "provided" is unsupported. Please add a --providedRuntime CLI option.');
-      }
-    }
-
-    if (!(serviceRuntime.startsWith('nodejs') || serviceRuntime.startsWith('python') || serviceRuntime.startsWith('ruby'))) {
-      this.printBlankLine();
-      this.serverlessLog(`Warning: found unsupported runtime '${serviceRuntime}'`);
-
-      return;
-    }
-
-    // for simple API Key authentication model
-    if (apiKeys) {
-      this.serverlessLog(`Key with token: ${this.options.apiKey}`);
-
-      if (this.options.noAuth) {
-        this.serverlessLog('Authorizers are turned off. You do not need to use x-api-key header.');
-      }
-      else {
-        this.serverlessLog('Remember to use x-api-key on the request headers');
-      }
-    }
-
-    Object.keys(this.service.functions).forEach(key => {
-
-      const fun = this.service.getFunction(key);
-      const funName = key;
-      const servicePath = path.join(this.serverless.config.servicePath, this.options.location);
-      const funOptions = functionHelper.getFunctionOptions(fun, key, servicePath, serviceRuntime);
-
-      debugLog(`funOptions ${JSON.stringify(funOptions, null, 2)} `);
-      this.printBlankLine();
-      debugLog(funName, 'runtime', serviceRuntime);
-      this.serverlessLog(`Routes for ${funName}:`);
-
-      if (!fun.events) {
-        fun.events = [];
-      }
-
-      // Add proxy for lamda invoke
-      fun.events.push({
-        http: {
-          method: 'POST',
-          path: `{apiVersion}/functions/${fun.name}/invocations`,
-          integration: 'lambda',
-          request: {
-            template: {
-              // AWS SDK for NodeJS specifies as 'binary/octet-stream' not 'application/json'
-              'binary/octet-stream': '$input.body',
-            },
-          },
-          response: {
-            headers: {
-              'Content-Type': 'application/json',
-            },
-          },
-        },
-      });
-      // Adds a route for each http endpoint
-      // eslint-disable-next-line
-      fun.events.forEach(event => {
-        if (event.websocket) {
-          this._createWsAction(fun, funName, servicePath, funOptions, event);
-
-          return;
-        }
-        if (!event.http) return;
-
-        // Handle Simple http setup, ex. - http: GET users/index
-        if (typeof event.http === 'string') {
-          const [method, path] = event.http.split(' ');
-          event.http = { method, path };
-        }
-
-        // generate an enpoint via the endpoint class
-        const endpoint = new Endpoint(event.http, funOptions).generate();
-
-        const integration = endpoint.integration || 'lambda-proxy';
-        const requestBodyValidationModel = (['lambda', 'lambda-proxy'].includes(integration)
-          ? requestBodyValidator.getModel(this.service.custom, event.http, this.serverlessLog)
-          : null);
-        const epath = endpoint.path;
-        const method = endpoint.method.toUpperCase();
-        const requestTemplates = endpoint.requestTemplates;
-
-        // Prefix must start and end with '/' BUT path must not end with '/'
-        let fullPath = this.options.prefix + (epath.startsWith('/') ? epath.slice(1) : epath);
-        if (fullPath !== '/' && fullPath.endsWith('/')) fullPath = fullPath.slice(0, -1);
-        fullPath = fullPath.replace(/\+}/g, '*}');
-
-        if (event.http.private) {
-          protectedRoutes.push(`${method}#${fullPath}`);
-        }
-
-        this.serverlessLog(`${method} ${fullPath}${requestBodyValidationModel && !this.options.disableModelValidation ? ` - request body will be validated against ${requestBodyValidationModel.name}` : ''}`);
-
-        // If the endpoint has an authorization function, create an authStrategy for the route
-        const authStrategyName = this.options.noAuth ? null : this._configureAuthorization(endpoint, funName, method, epath, servicePath, serviceRuntime);
-
-        let cors = null;
-        if (endpoint.cors) {
-          cors = {
-            origin: endpoint.cors.origins || this.options.corsConfig.origin,
-            headers: endpoint.cors.headers || this.options.corsConfig.headers,
-            credentials: endpoint.cors.credentials || this.options.corsConfig.credentials,
-            exposedHeaders: this.options.corsConfig.exposedHeaders,
-          };
-        }
-
-        // Route creation
-        const routeMethod = method === 'ANY' ? '*' : method;
-
-        const state = this.options.disableCookieValidation ? {
-          parse: false,
-          failAction: 'ignore',
-        } : {
-          parse: true,
-          failAction: 'error',
-        };
-
-        const routeConfig = {
-          cors,
-          auth: authStrategyName,
-          timeout: { socket: false },
-          state,
-        };
-
-        // skip HEAD routes as hapi will fail with 'Method name not allowed: HEAD ...'
-        // for more details, check https://github.com/dherault/serverless-offline/issues/204
-        if (routeMethod === 'HEAD') {
-          this.serverlessLog('HEAD method event detected. Skipping HAPI server route mapping ...');
-
-          return;
-        }
-
-        if (routeMethod !== 'HEAD' && routeMethod !== 'GET') {
-          // maxBytes: Increase request size from 1MB default limit to 10MB.
-          // Cf AWS API GW payload limits.
-          routeConfig.payload = { parse: false, maxBytes: 1024 * 1024 * 10 };
-        }
-
-        this.server.route({
-          method: routeMethod,
-          path: fullPath,
-          config: routeConfig,
-          handler: (request, h) => { // Here we go
-            // Payload processing
-            const encoding = detectEncoding(request);
-
-            request.payload = request.payload && request.payload.toString(encoding);
-            request.rawPayload = request.payload;
-
-            // Headers processing
-            // Hapi lowercases the headers whereas AWS does not
-            // so we recreate a custom headers object from the raw request
-            const headersArray = request.raw.req.rawHeaders;
-
-            // During tests, `server.inject` uses *shot*, a package
-            // for performing injections that does not entirely mimick
-            // Hapi's usual request object. rawHeaders are then missing
-            // Hence the fallback for testing
-
-            // Normal usage
-            if (headersArray) {
-              request.unprocessedHeaders = {};
-              request.multiValueHeaders = {};
-
-              for (let i = 0; i < headersArray.length; i += 2) {
-                request.unprocessedHeaders[headersArray[i]] = headersArray[i + 1];
-                request.multiValueHeaders[headersArray[i]] = (request.multiValueHeaders[headersArray[i]] || []).concat(headersArray[i + 1]);
-              }
-            }
-            // Lib testing
-            else {
-              request.unprocessedHeaders = request.headers;
-            }
-
-            // Incomming request message
-            this.printBlankLine();
-            this.serverlessLog(`${method} ${request.path} (λ: ${funName})`);
-
-            // Check for APIKey
-            if ((protectedRoutes.includes(`${routeMethod}#${fullPath}`) || protectedRoutes.includes(`ANY#${fullPath}`)) && !this.options.noAuth) {
-              const errorResponse = () => h.response({ message: 'Forbidden' }).code(403).type('application/json').header('x-amzn-ErrorType', 'ForbiddenException');
-
-              if ('x-api-key' in request.headers) {
-                const requestToken = request.headers['x-api-key'];
-                if (requestToken !== this.options.apiKey) {
-                  debugLog(`Method ${method} of function ${funName} token ${requestToken} not valid`);
-
-                  return errorResponse();
-                }
-              }
-              else if (request.auth && request.auth.credentials && 'usageIdentifierKey' in request.auth.credentials) {
-                const usageIdentifierKey = request.auth.credentials.usageIdentifierKey;
-                if (usageIdentifierKey !== this.options.apiKey) {
-                  debugLog(`Method ${method} of function ${funName} token ${usageIdentifierKey} not valid`);
-
-                  return errorResponse();
-                }
-              }
-              else {
-                debugLog(`Missing x-api-key on private function ${funName}`);
-
-                return errorResponse();
-              }
-            }
-            // Shared mutable state is the root of all evil they say
-            const requestId = getUniqueId();
-            this.requests[requestId] = { done: false };
-            this.currentRequestId = requestId;
-
-            const response = h.response();
-            const contentType = request.mime || defaultContentType;
-
-            // default request template to '' if we don't have a definition pushed in from serverless or endpoint
-            const requestTemplate = typeof requestTemplates !== 'undefined' && integration === 'lambda' ? requestTemplates[contentType] : '';
-
-            // https://hapijs.com/api#route-configuration doesn't seem to support selectively parsing
-            // so we have to do it ourselves
-            const contentTypesThatRequirePayloadParsing = ['application/json', 'application/vnd.api+json'];
-            if (contentTypesThatRequirePayloadParsing.includes(contentType) && request.payload && request.payload.length > 1) {
-              try {
-                if (!request.payload || request.payload.length < 1) {
-                  request.payload = '{}';
-                }
-
-                request.payload = JSON.parse(request.payload);
-              }
-              catch (err) {
-                debugLog('error in converting request.payload to JSON:', err);
-              }
-            }
-
-            debugLog('requestId:', requestId);
-            debugLog('contentType:', contentType);
-            debugLog('requestTemplate:', requestTemplate);
-            debugLog('payload:', request.payload);
-
-            /* HANDLER LAZY LOADING */
-
-            let userHandler; // The lambda function
-            Object.assign(process.env, this.originalEnvironment);
-
-            try {
-              if (this.options.noEnvironment) {
-                // This evict errors in server when we use aws services like ssm
-                const baseEnvironment = {
-                  AWS_REGION: 'dev',
-                };
-                if (!process.env.AWS_PROFILE) {
-                  baseEnvironment.AWS_ACCESS_KEY_ID = 'dev';
-                  baseEnvironment.AWS_SECRET_ACCESS_KEY = 'dev';
-                }
-
-                process.env = Object.assign(baseEnvironment, process.env);
-              }
-              else {
-                Object.assign(
-                  process.env,
-                  { AWS_REGION: this.service.provider.region },
-                  this.service.provider.environment,
-                  this.service.functions[key].environment
-                );
-              }
-              process.env._HANDLER = fun.handler;
-              userHandler = functionHelper.createHandler(funOptions, this.options);
-            }
-            catch (err) {
-              return this._reply500(response, `Error while loading ${funName}`, err);
-            }
-
-            /* REQUEST TEMPLATE PROCESSING (event population) */
-
-            let event = {};
-
-            if (integration === 'lambda') {
-              if (requestTemplate) {
-                try {
-                  debugLog('_____ REQUEST TEMPLATE PROCESSING _____');
-                  // Velocity templating language parsing
-                  const velocityContext = createVelocityContext(request, this.velocityContextOptions, request.payload || {});
-                  event = renderVelocityTemplateObject(requestTemplate, velocityContext);
-                }
-                catch (err) {
-                  return this._reply500(response, `Error while parsing template "${contentType}" for ${funName}`, err);
-                }
-              }
-              else if (typeof request.payload === 'object') {
-                event = request.payload || {};
-              }
-            }
-            else if (integration === 'lambda-proxy') {
-              event = createLambdaProxyContext(request, this.options, this.velocityContextOptions.stageVariables);
-            }
-
-            event.isOffline = true;
-
-            if (this.service.custom && this.service.custom.stageVariables) {
-              event.stageVariables = this.service.custom.stageVariables;
-            }
-            else if (integration !== 'lambda-proxy') {
-              event.stageVariables = {};
-            }
-
-            debugLog('event:', event);
-
-            return new Promise(resolve => {
-              // We create the context, its callback (context.done/succeed/fail) will send the HTTP response
-              const lambdaContext = createLambdaContext(fun, this.service.provider, (err, data, fromPromise) => {
-                // Everything in this block happens once the lambda function has resolved
-                debugLog('_____ HANDLER RESOLVED _____');
-
-                // User should not call context.done twice
-                if (!this.requests[requestId] || this.requests[requestId].done) {
-                  this.printBlankLine();
-                  const warning = fromPromise
-                    ? `Warning: handler '${funName}' returned a promise and also uses a callback!\nThis is problematic and might cause issues in your lambda.`
-                    : `Warning: context.done called twice within handler '${funName}'!`;
-                  this.serverlessLog(warning);
-                  debugLog('requestId:', requestId);
-
-                  return;
-                }
-
-                this.requests[requestId].done = true;
-
-                let result = data;
-                let responseName = 'default';
-                const { contentHandling, responseContentType } = endpoint;
-
-                /* RESPONSE SELECTION (among endpoint's possible responses) */
-
-                // Failure handling
-                let errorStatusCode = 0;
-                if (err) {
-                  // Since the --useSeparateProcesses option loads the handler in
-                  // a separate process and serverless-offline communicates with it
-                  // over IPC, we are unable to catch JavaScript unhandledException errors
-                  // when the handler code contains bad JavaScript. Instead, we "catch"
-                  // it here and reply in the same way that we would have above when
-                  // we lazy-load the non-IPC handler function.
-                  if (this.options.useSeparateProcesses && err.ipcException) {
-                    return resolve(this._reply500(response, `Error while loading ${funName}`, err));
-                  }
-
-                  const errorMessage = (err.message || err).toString();
-
-                  const re = /\[(\d{3})]/;
-                  const found = errorMessage.match(re);
-                  if (found && found.length > 1) {
-                    errorStatusCode = found[1];
-                  }
-                  else {
-                    errorStatusCode = '500';
-                  }
-
-                  // Mocks Lambda errors
-                  result = {
-                    errorMessage,
-                    errorType: err.constructor.name,
-                    stackTrace: this._getArrayStackTrace(err.stack),
-                  };
-
-                  this.serverlessLog(`Failure: ${errorMessage}`);
-
-                  if (result.stackTrace) {
-                    debugLog(result.stackTrace.join('\n  '));
-                  }
-
-                  for (const key in endpoint.responses) {
-                    if (key !== 'default' && errorMessage.match(`^${endpoint.responses[key].selectionPattern || key}$`)) {
-                      responseName = key;
-                      break;
-                    }
-                  }
-                }
-
-                debugLog(`Using response '${responseName}'`);
-                const chosenResponse = endpoint.responses[responseName];
-
-                /* RESPONSE PARAMETERS PROCCESSING */
-
-                const responseParameters = chosenResponse.responseParameters;
-
-                if (responseParameters) {
-
-                  const responseParametersKeys = Object.keys(responseParameters);
-
-                  debugLog('_____ RESPONSE PARAMETERS PROCCESSING _____');
-                  debugLog(`Found ${responseParametersKeys.length} responseParameters for '${responseName}' response`);
-
-                  // responseParameters use the following shape: "key": "value"
-                  Object.entries(responseParametersKeys).forEach(([key, value]) => {
-
-                    const keyArray = key.split('.'); // eg: "method.response.header.location"
-                    const valueArray = value.split('.'); // eg: "integration.response.body.redirect.url"
-
-                    debugLog(`Processing responseParameter "${key}": "${value}"`);
-
-                    // For now the plugin only supports modifying headers
-                    if (key.startsWith('method.response.header') && keyArray[3]) {
-
-                      const headerName = keyArray.slice(3).join('.');
-                      let headerValue;
-                      debugLog('Found header in left-hand:', headerName);
-
-                      if (value.startsWith('integration.response')) {
-                        if (valueArray[2] === 'body') {
-
-                          debugLog('Found body in right-hand');
-                          headerValue = (valueArray[3] ? jsonPath(result, valueArray.slice(3).join('.')) : result).toString();
-
-                        }
-                        else {
-                          this.printBlankLine();
-                          this.serverlessLog(`Warning: while processing responseParameter "${key}": "${value}"`);
-                          this.serverlessLog(`Offline plugin only supports "integration.response.body[.JSON_path]" right-hand responseParameter. Found "${value}" instead. Skipping.`);
-                          this.logPluginIssue();
-                          this.printBlankLine();
-                        }
-                      }
-                      else {
-                        headerValue = value.match(/^'.*'$/) ? value.slice(1, -1) : value; // See #34
-                      }
-                      // Applies the header;
-                      debugLog(`Will assign "${headerValue}" to header "${headerName}"`);
-                      response.header(headerName, headerValue);
-
-                    }
-                    else {
-                      this.printBlankLine();
-                      this.serverlessLog(`Warning: while processing responseParameter "${key}": "${value}"`);
-                      this.serverlessLog(`Offline plugin only supports "method.response.header.PARAM_NAME" left-hand responseParameter. Found "${key}" instead. Skipping.`);
-                      this.logPluginIssue();
-                      this.printBlankLine();
-                    }
-                  });
-                }
-
-                let statusCode = 200;
-
-                if (integration === 'lambda') {
-
-                  const endpointResponseHeaders = (endpoint.response && endpoint.response.headers) || {};
-
-                  Object.entries(endpointResponseHeaders)
-                    .filter(([, value]) => typeof value === 'string' && /^'.*?'$/.test(value))
-                    .forEach(([key, value]) => response.header(key, value.slice(1, -1)));
-
-                  /* LAMBDA INTEGRATION RESPONSE TEMPLATE PROCCESSING */
-
-                  // If there is a responseTemplate, we apply it to the result
-                  const { responseTemplates } = chosenResponse;
-
-                  if (typeof responseTemplates === 'object') {
-                    const responseTemplatesKeys = Object.keys(responseTemplates);
-
-                    if (responseTemplatesKeys.length) {
-
-                      // BAD IMPLEMENTATION: first key in responseTemplates
-                      const responseTemplate = responseTemplates[responseContentType];
-
-                      if (responseTemplate && responseTemplate !== '\n') {
-
-                        debugLog('_____ RESPONSE TEMPLATE PROCCESSING _____');
-                        debugLog(`Using responseTemplate '${responseContentType}'`);
-
-                        try {
-                          const reponseContext = createVelocityContext(request, this.velocityContextOptions, result);
-                          result = renderVelocityTemplateObject({ root: responseTemplate }, reponseContext).root;
-                        }
-                        catch (error) {
-                          this.serverlessLog(`Error while parsing responseTemplate '${responseContentType}' for lambda ${funName}:`);
-                          console.log(error.stack);
-                        }
-                      }
-                    }
-                  }
-
-                  /* LAMBDA INTEGRATION HAPIJS RESPONSE CONFIGURATION */
-
-                  statusCode = errorStatusCode !== 0 ? errorStatusCode : (chosenResponse.statusCode || 200);
-
-                  if (!chosenResponse.statusCode) {
-                    this.printBlankLine();
-                    this.serverlessLog(`Warning: No statusCode found for response "${responseName}".`);
-                  }
-
-                  response.header('Content-Type', responseContentType, {
-                    override: false, // Maybe a responseParameter set it already. See #34
-                  });
-
-                  response.statusCode = statusCode;
-
-                  if (contentHandling === 'CONVERT_TO_BINARY') {
-                    response.encoding = 'binary';
-                    response.source = Buffer.from(result, 'base64');
-                    response.variety = 'buffer';
-                  }
-                  else {
-                    if (result && result.body && typeof result.body !== 'string') {
-                      return this._reply500(response, 'According to the API Gateway specs, the body content must be stringified. Check your Lambda response and make sure you are invoking JSON.stringify(YOUR_CONTENT) on your body object', {});
-                    }
-                    response.source = result;
-                  }
-                }
-                else if (integration === 'lambda-proxy') {
-
-                  /* LAMBDA PROXY INTEGRATION HAPIJS RESPONSE CONFIGURATION */
-
-                  response.statusCode = statusCode = result.statusCode || 200;
-
-                  const headers = {};
-                  if (result.headers) {
-                    Object.keys(result.headers).forEach(header => {
-                      headers[header] = (headers[header] || []).concat(result.headers[header]);
-                    });
-                  }
-                  if (result.multiValueHeaders) {
-                    Object.keys(result.multiValueHeaders).forEach(header => {
-                      headers[header] = (headers[header] || []).concat(result.multiValueHeaders[header]);
-                    });
-                  }
-
-                  debugLog('headers', headers);
-
-                  Object.keys(headers).forEach(header => {
-                    if (header.toLowerCase() === 'set-cookie') {
-                      headers[header].forEach(headerValue => {
-                        const cookieName = headerValue.slice(0, headerValue.indexOf('='));
-                        const cookieValue = headerValue.slice(headerValue.indexOf('=') + 1);
-                        h.state(cookieName, cookieValue, { encoding: 'none', strictHeader: false });
-                      });
-                    }
-                    else {
-                      headers[header].forEach(headerValue => {
-                        // it looks like Hapi doesn't support multiple headers with the same name,
-                        // appending values is the closest we can come to the AWS behavior.
-                        response.header(header, headerValue, { append: true });
-                      });
-                    }
-                  });
-
-                  response.header('Content-Type', 'application/json', { override: false, duplicate: false });
-
-                  if (typeof result.body !== 'undefined') {
-                    if (result.isBase64Encoded) {
-                      response.encoding = 'binary';
-                      response.source = Buffer.from(result.body, 'base64');
-                      response.variety = 'buffer';
-                    }
-                    else {
-                      if (result.body && typeof result.body !== 'string') {
-                        return this._reply500(response, 'According to the API Gateway specs, the body content must be stringified. Check your Lambda response and make sure you are invoking JSON.stringify(YOUR_CONTENT) on your body object', {});
-                      }
-                      response.source = result.body;
-                    }
-                  }
-                }
-
-                // Log response
-                let whatToLog = result;
-
-                try {
-                  whatToLog = JSON.stringify(result);
-                }
-                catch (error) {
-                  // nothing
-                }
-                finally {
-                  if (this.options.printOutput) this.serverlessLog(err ? `Replying ${statusCode}` : `[${statusCode}] ${whatToLog}`);
-                  debugLog('requestId:', requestId);
-                }
-
-                // Bon voyage!
-                resolve(response);
-              });
-
-              // Now we are outside of createLambdaContext, so this happens before the handler gets called:
-
-              // We cannot use Hapijs's timeout feature because the logic above can take a significant time, so we implement it ourselves
-              this.requests[requestId].timeout = this.options.noTimeout ? null : setTimeout(
-                this._replyTimeout.bind(this, response, resolve, funName, funOptions.funTimeout, requestId),
-                funOptions.funTimeout
-              );
-
-              // If request body validation is enabled, validate body against the request model.
-              if (requestBodyValidationModel && !this.options.disableModelValidation) {
-                try {
-                  requestBodyValidator.validate(requestBodyValidationModel, event.body);
-                }
-                catch (error) {
-                  // When request body validation fails, APIG will return back 400 as detailed in:
-                  // https://docs.aws.amazon.com/apigateway/latest/developerguide/api-gateway-method-request-validation.html
-                  return resolve(this._replyError(400, response, `Invalid request body for '${funName}' handler`, error));
-                }
-              }
-
-              // Finally we call the handler
-              debugLog('_____ CALLING HANDLER _____');
-
-              const cleanup = () => {
-                this._clearTimeout(requestId);
-                delete this.requests[requestId];
-              };
-
-              let x;
-
-              if (this.options.showDuration) {
-                performance.mark(`${requestId}-start`);
-
-                const obs = new PerformanceObserver(list => {
-                  for (const entry of list.getEntries()) {
-                    this.serverlessLog(`Duration ${entry.duration.toFixed(2)} ms (λ: ${entry.name})`);
-                  }
-
-                  obs.disconnect();
-                });
-
-                obs.observe({ entryTypes: ['measure'] });
-              }
-
-              try {
-                x = userHandler(event, lambdaContext, (err, result) => {
-                  setTimeout(cleanup, 0);
-
-                  if (this.options.showDuration) {
-                    performance.mark(`${requestId}-end`);
-                    performance.measure(funName, `${requestId}-start`, `${requestId}-end`);
-                  }
-
-                  return lambdaContext.done(err, result);
-                });
-
-                // Promise support
-                if (!this.requests[requestId].done) {
-                  if (x && typeof x.then === 'function') {
-                    x.then(lambdaContext.succeed).catch(lambdaContext.fail).then(cleanup, cleanup);
-                  }
-                  else if (x instanceof Error) {
-                    lambdaContext.fail(x);
-                  }
-                }
-              }
-              catch (error) {
-                cleanup();
-
-                return resolve(this._reply500(response, `Uncaught error in your '${funName}' handler`, error));
-              }
-            });
-          },
-        });
-      });
-    });
-  }
-
-  _extractAuthFunctionName(endpoint) {
-    const result = authFunctionNameExtractor(endpoint, this.serverlessLog);
-
-    return result.unsupportedAuth ? null : result.authorizerName;
-  }
-
-  _configureAuthorization(endpoint, funName, method, epath, servicePath, serviceRuntime) {
-    if (!endpoint.authorizer) {
-      return null;
-    }
-
-    const authFunctionName = this._extractAuthFunctionName(endpoint);
-
-    if (!authFunctionName) {
-      return null;
-    }
-
-    this.serverlessLog(`Configuring Authorization: ${endpoint.path} ${authFunctionName}`);
-
-    const authFunction = this.service.getFunction(authFunctionName);
-
-    if (!authFunction) return this.serverlessLog(`WARNING: Authorization function ${authFunctionName} does not exist`);
-
-    const authorizerOptions = {
-      resultTtlInSeconds: '300',
-      identitySource: 'method.request.header.Authorization',
-      identityValidationExpression: '(.*)',
-    };
-
-    if (typeof endpoint.authorizer === 'string') {
-      authorizerOptions.name = authFunctionName;
-    }
-    else {
-      Object.assign(authorizerOptions, endpoint.authorizer);
-    }
-
-    // Create a unique scheme per endpoint
-    // This allows the methodArn on the event property to be set appropriately
-    const authKey = `${funName}-${authFunctionName}-${method}-${epath}`;
-    const authSchemeName = `scheme-${authKey}`;
-    const authStrategyName = `strategy-${authKey}`; // set strategy name for the route config
-
-    debugLog(`Creating Authorization scheme for ${authKey}`);
-
-    // Create the Auth Scheme for the endpoint
-    const scheme = createAuthScheme(
-      authFunction,
-      authorizerOptions,
-      authFunctionName,
-      epath,
-      this.options,
-      this.serverlessLog,
-      servicePath,
-      serviceRuntime,
-      this.serverless
-    );
-
-    // Set the auth scheme and strategy on the server
-    this.server.auth.scheme(authSchemeName, scheme);
-    this.server.auth.strategy(authStrategyName, authSchemeName);
-
-    return authStrategyName;
-  }
-
-  // All done, we can listen to incomming requests
-  async _listen() {
-    try {
-      await this.server.start();
-    }
-    catch (e) {
-      console.error(`Unexpected error while starting serverless-offline server on port ${this.options.port}:`, e);
-      process.exit(1);
-    }
-
-    this.printBlankLine();
-    this.serverlessLog(`Offline listening on http${this.options.httpsProtocol ? 's' : ''}://${this.options.host}:${this.options.port}`);
-
-    try {
-      await this.wsServer.start();
-    }
-    catch (e) {
-      console.error(`Unexpected error while starting serverless-offline websocket server on port ${this.options.wsPort}:`, e);
-      process.exit(1);
-    }
-
-    this.printBlankLine();
-    this.serverlessLog(`Offline listening on ws${this.options.httpsProtocol ? 's' : ''}://${this.options.host}:${this.options.wsPort}`);
-
-    this.printBlankLine();
-    this.serverlessLog(`Offline listening on http${this.options.httpsProtocol ? 's' : ''}://${this.options.host}:${this.options.wsPort}/@connections/{connectionId}`);
-  }
-
   end() {
     this.serverlessLog('Halting offline server');
     functionHelper.cleanup();
-    this.server.stop({ timeout: 5000 })
+    this.apiGateway.server.stop({ timeout: 5000 })
       .then(() => process.exit(this.exitCode));
   }
-
-  // Bad news
-  _replyError(responseCode, response, message, err) {
-    const stackTrace = this._getArrayStackTrace(err.stack);
-
-    this.serverlessLog(message);
-    if (stackTrace && stackTrace.length > 0) {
-      console.log(stackTrace);
-    }
-    else {
-      console.log(err);
-    }
-
-    response.header('Content-Type', 'application/json');
-
-    /* eslint-disable no-param-reassign */
-    response.statusCode = responseCode;
-    response.source = {
-      errorMessage: message,
-      errorType: err.constructor.name,
-      stackTrace,
-      offlineInfo: 'If you believe this is an issue with the plugin please submit it, thanks. https://github.com/dherault/serverless-offline/issues',
-    };
-    /* eslint-enable no-param-reassign */
-    this.serverlessLog('Replying error in handler');
-
-    return response;
-  }
-
-  _reply500(response, message, err) {
-    // APIG replies 200 by default on failures
-    return this._replyError(200, response, message, err);
-  }
-
-  _replyTimeout(response, resolve, funName, funTimeout, requestId) {
-    if (this.currentRequestId !== requestId) return;
-
-    this.serverlessLog(`Replying timeout after ${funTimeout}ms`);
-    /* eslint-disable no-param-reassign */
-    response.statusCode = 503;
-    response.source = `[Serverless-Offline] Your λ handler '${funName}' timed out after ${funTimeout}ms.`;
-    /* eslint-enable no-param-reassign */
-    resolve(response);
-  }
-
-  _clearTimeout(requestId) {
-    const { timeout } = this.requests[requestId] || {};
-    clearTimeout(timeout);
-  }
-
-  _createResourceRoutes() {
-    if (!this.options.resourceRoutes) return true;
-    const resourceRoutesOptions = this.options.resourceRoutes;
-    const resourceRoutes = parseResources(this.service.resources);
-
-    if (!resourceRoutes || !Object.keys(resourceRoutes).length) return true;
-
-    this.printBlankLine();
-    this.serverlessLog('Routes defined in resources:');
-
-    Object.entries(resourceRoutes).forEach(([methodId, resourceRoutesObj]) => {
-      const { isProxy, method, path, pathResource, proxyUri } = resourceRoutesObj;
-
-      if (!isProxy) {
-        return this.serverlessLog(`WARNING: Only HTTP_PROXY is supported. Path '${pathResource}' is ignored.`);
-      }
-      if (!path) {
-        return this.serverlessLog(`WARNING: Could not resolve path for '${methodId}'.`);
-      }
-
-      let fullPath = this.options.prefix + (pathResource.startsWith('/') ? pathResource.slice(1) : pathResource);
-      if (fullPath !== '/' && fullPath.endsWith('/')) fullPath = fullPath.slice(0, -1);
-      fullPath = fullPath.replace(/\+}/g, '*}');
-
-      const proxyUriOverwrite = resourceRoutesOptions[methodId] || {};
-      const proxyUriInUse = proxyUriOverwrite.Uri || proxyUri;
-
-      if (!proxyUriInUse) {
-        return this.serverlessLog(`WARNING: Could not load Proxy Uri for '${methodId}'`);
-      }
-
-      const routeMethod = method === 'ANY' ? '*' : method;
-      const routeConfig = { cors: this.options.corsConfig };
-
-      // skip HEAD routes as hapi will fail with 'Method name not allowed: HEAD ...'
-      // for more details, check https://github.com/dherault/serverless-offline/issues/204
-      if (routeMethod === 'HEAD') {
-        this.serverlessLog('HEAD method event detected. Skipping HAPI server route mapping ...');
-
-        return;
-      }
-
-      if (routeMethod !== 'HEAD' && routeMethod !== 'GET') {
-        routeConfig.payload = { parse: false };
-      }
-
-      this.serverlessLog(`${method} ${fullPath} -> ${proxyUriInUse}`);
-      this.server.route({
-        method: routeMethod,
-        path: fullPath,
-        config: routeConfig,
-        handler: (request, h) => {
-          const { params } = request;
-          let resultUri = proxyUriInUse;
-
-          Object.entries(params).forEach(([key, value]) => {
-            resultUri = resultUri.replace(`{${key}}`, value);
-          });
-
-          if (request.url.search !== null) {
-            resultUri += request.url.search; // search is empty string by default
-          }
-
-          this.serverlessLog(`PROXY ${request.method} ${request.url.path} -> ${resultUri}`);
-
-          return h.proxy({ uri: resultUri, passThrough: true });
-        },
-      });
-    });
-  }
-
-  _create404Route() {
-    // If a {proxy+} route exists, don't conflict with it
-    if (this.server.match('*', '/{p*}')) return;
-
-    this.server.route({
-      method: '*',
-      path: '/{p*}',
-      config: { cors: this.options.corsConfig },
-      handler: (request, h) => {
-        const response = h.response({
-          statusCode: 404,
-          error: 'Serverless-offline: route not found.',
-          currentRoute: `${request.method} - ${request.path}`,
-          existingRoutes: this.server.table()
-            .filter(route => route.path !== '/{p*}') // Exclude this (404) route
-            .sort((a, b) => a.path <= b.path ? -1 : 1) // Sort by path
-            .map(route => `${route.method} - ${route.path}`), // Human-friendly result
-        });
-        response.statusCode = 404;
-
-        return response;
-      },
-    });
-  }
-
-  _getArrayStackTrace(stack) {
-    if (!stack) return null;
-
-    const splittedStack = stack.split('\n');
-
-    return splittedStack.slice(0, splittedStack.findIndex(item => item.match(/server.route.handler.createLambdaContext/))).map(line => line.trim());
-  }
-
-  _logAndExit() {
-    // eslint-disable-next-line
-    console.log.apply(null, arguments);
-    process.exit(0);
-  }
-}
+};
 
 // Serverless exits with code 1 when a promise rejection is unhandled. Not AWS.
 // Users can still use their own unhandledRejection event though.
 process.removeAllListeners('unhandledRejection');
-
-module.exports = Offline;
